src/main/java/accio/hogsmeade/store/client/member/validator/MemberValidator.java
package accio.hogsmeade.store.client.member.validator;

import accio.hogsmeade.store.client.member.Member;
import accio.hogsmeade.store.client.member.repository.MemberRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.context.annotation.Configuration;

import java.util.NoSuchElementException;

@Configuration
@RequiredArgsConstructor
public class MemberValidator {

    private final MemberRepository memberRepository;

    public Member findById(Long id) {
        return memberRepository.findById(id)
                .orElseThrow(NoSuchElementException::new);
    }

    public Member findByLoginId(String loginId) {
        return memberRepository.findByLoginId(loginId)
                .orElseThrow(NoSuchElementException::new);
    }

    public Member findByTel(String tel) {
        return memberRepository.findByTel(tel)
                .orElseThrow(NoSuchElementException::new);
    }
}
